+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - support parsing unknown artifacts and army creatures (pull request #23)
 - add missing support to Heroes Chronicles savefiles for Shadow of Death creatures and artifacts and spells
 - improve parsing map name and description from savefile
+- improve retaining equipment and inventory original bytes on no change
 - write savefiles without filename in ZIP header, just as the game does
 - fix error on refilling spell points if hero has Intelligence skill